requirements.txt
fastapi
uvicorn[standard]
boto3
python-multipart
python-dotenv
databases
aiosqlite
sqlalchemy
httpx
gunicorn
aiofiles
passlib[bcrypt]
python-jose[cryptography]
aiosmtplib
email-validator
Authlib
alembic
pydantic[email]
argon2-cffi